requirements.txt
# Newsletter RAG Dependencies
# ========================================

# Core Framework
streamlit>=1.28.0

# Gmail API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# AI & Embeddings
google-generativeai>=0.3.0

# HTML Cleaning
trafilatura>=1.6.0

# Vector Database
chromadb>=0.4.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
